Fix list of default break chars in man page
There are two double quotes in the list of default break chars to keep
Emacs’s syntax coloring happy, but they both appear in the man page too,
which is confusing. Keep Emacs fairly happy with an extra quote in a comment
at the end of the line, but avoid the bogus output.